requirements.txt
asgiref==3.8.1
azure-core==1.35.0
azure-storage-blob==12.25.1
boto3==1.21.4
botocore==1.24.5
certifi==2025.6.15
cffi==1.17.1
charset-normalizer==3.4.2
cryptography==45.0.5
Django==4.2.24
django-cors-headers==4.7.0
django-debug-toolbar==5.2.0
django-filter==25.1
django-multitenant==4.1.1
django-oauth-toolkit==3.0.1
django-rest-authtoken==2.1.4
django-storages==1.14.6
djangorestframework==3.14.0
djangorestframework-simplejwt==5.3.0
et_xmlfile==2.0.0
idna==3.10
isodate==0.7.2
jmespath==0.10.0
jwcrypto==1.5.6
numpy==2.3.1
oauthlib==3.3.1
openpyxl==3.1.5
pandas==2.3.0
pillow==10.3.0
psycopg2-binary==2.9.10
pycparser==2.22
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.1.1
pytz==2025.2
requests==2.32.4
s3transfer==0.5.2
sentry-sdk==2.32.0
setuptools==80.9.0
six==1.17.0
sqlparse==0.5.3
typesense==1.1.1
typing_extensions==4.14.1
tzdata==2025.2
urllib3==1.26.20